# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server-side api of the Octane.",
        "title": "OctaneServer",
        "contact": {
            "name": "GitHub",
            "url": "https://github.com/Team-Kamo/server"
        },
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "get health status of the api server",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Get health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Health"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            }
        },
        "/room": {
            "post": {
                "description": "Create new room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create room",
                "parameters": [
                    {
                        "description": "Room create request",
                        "name": "RoomCreateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.RoomCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.RoomCreate"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            }
        },
        "/room/{id}": {
            "get": {
                "description": "Get status of the room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "Get room status",
                "parameters": [
                    {
                        "description": "RoomCreateRequest",
                        "name": "RoomCreateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.RoomCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Room"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Connect device to the room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "room"
                ],
                "summary": "Connect to the room",
                "parameters": [
                    {
                        "description": "Room connect request",
                        "name": "RoomConnectRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.RoomConnectRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "room"
                ],
                "summary": "Delete room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            }
        },
        "/room/{id}/content": {
            "get": {
                "description": "Get content to the room",
                "tags": [
                    "content"
                ],
                "summary": "Get room content",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Set content to the room",
                "tags": [
                    "content"
                ],
                "summary": "Set room content",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete content to the room",
                "tags": [
                    "content"
                ],
                "summary": "Delete room content",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            }
        },
        "/room/{id}/status": {
            "get": {
                "description": "Get content status of the room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Get content status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Status"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Set content status of the room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Set content status",
                "parameters": [
                    {
                        "description": "Room status",
                        "name": "Status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Status"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete content status of the room",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Delete content status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.Device": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "data.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "data.Health": {
            "type": "object",
            "properties": {
                "health": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "data.Room": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.Device"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "data.RoomConnectRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "data.RoomCreate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "data.RoomCreateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "data.Status": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "mime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "X-Octane-API-Token": {
            "type": "apiKey",
            "name": "X-Octane-API-Token",
            "in": "header"
        }
    }
}